requirements.txt
cassandra-driver
pyyaml